More stable test setup/teardown
