+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/polyfill/web/TimersTest.java
@@ -1,28 +1,26 @@
 package org.enso.ydoc.polyfill.web;
 
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
+import org.enso.ydoc.polyfill.ExecutorSetup;
 import org.graalvm.polyglot.Context;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class TimersTest {
+public class TimersTest extends ExecutorSetup {
 
   private static final Consumer<Object> NULL_CONSUMER = v -> {};
 
   private Context context;
-  private ExecutorService executor;
 
   public TimersTest() {}
 
   @Before
   public void setup() throws Exception {
-    executor = Executors.newSingleThreadExecutor();
+    super.setup();
     var timers = new Timers(executor);
     var contextBuilder = WebEnvironment.createContext();
 
@@ -38,8 +36,8 @@ public class TimersTest {
   }
 
   @After
-  public void tearDown() {
-    executor.shutdownNow();
+  public void tearDown() throws InterruptedException {
+    super.tearDown();
     context.close();
   }
