+++ lib/java/ydoc-server/src/test/java/org/enso/ydoc/polyfill/web/UtilTest.java
@@ -1,8 +1,7 @@
 package org.enso.ydoc.polyfill.web;
 
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import org.enso.ydoc.polyfill.ExecutorSetup;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.io.ByteSequence;
 import org.junit.After;
@@ -10,16 +9,15 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-public class UtilTest {
+public class UtilTest extends ExecutorSetup {
 
   private Context context;
-  private ExecutorService executor;
 
   public UtilTest() {}
 
   @Before
   public void setup() throws Exception {
-    executor = Executors.newSingleThreadExecutor();
+    super.setup();
     var encoding = new Util();
     var contextBuilder = WebEnvironment.createContext();
 
@@ -35,8 +33,8 @@ public class UtilTest {
   }
 
   @After
-  public void tearDown() {
-    executor.shutdownNow();
+  public void tearDown() throws InterruptedException {
+    super.tearDown();
     context.close();
   }
